feat(Gallery): added a navigator for the gallery
