+++ src/screens/Gallery/AlbumScreen.tsx
@@ -1,8 +1,9 @@
 import React, {useCallback, useEffect, useState} from 'react';
-import {listPhotosByAlbumId, PhotoResponse} from "../api/photo.api";
-import Photo from "../components/buttons/Photo";
-import {ItemContainerStyle} from "../components/containers/PrimaryContainer/style";
-import {ScrollItemContainerStyle} from "../components/containers/PrimaryScrollContainer/style";
+import {listPhotosByAlbumId, PhotoResponse} from "../../api/photo.api";
+import Photo from "../../components/buttons/Photo";
+import {ItemContainerStyle} from "../../components/containers/PrimaryContainer/style";
+import {ScrollItemContainerStyle} from "../../components/containers/PrimaryScrollContainer/style";
+import {ListRenderItemInfo} from "react-native";
 
 export default function AlbumScreen({route, navigation}: any) {
     const {albumId} = route.params;
@@ -15,11 +16,11 @@ export default function AlbumScreen({route, navigation}: any) {
         }
     }, [albumId]);
 
-    const renderItem: any = useCallback(({photo}: { photo: PhotoResponse }) => (
-        photo && <Photo
-            key={photo.id}
-            photo={photo}
-            onPress={() => navigation.navigate('PhotoDetails', {photo})}
+    const renderItem: any = useCallback(({item}: ListRenderItemInfo<PhotoResponse>) => (
+        <Photo
+            key={item.id}
+            photo={item}
+            onPress={() => navigation.navigate('Photo', {photo: item})}
         />
     ), []);
